backend/requirements.txt
Django==5.1.13
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.5.1
Pillow==10.4.0
python-decouple==3.8
psycopg2-binary==2.9.9
django-extensions==3.2.3
psutil==6.1.0
requests==2.32.4
dj-database-url==2.1.0